Separate raceResult logics from MatchManager for reuse
