+++ src/components/Footer/style.css
@@ -5,7 +5,7 @@
     width: 100%;
     height: 60px;
     line-height: 60px;
-    text-align: right;
+    text-align: center;
 }
 
 .copyRight {
